refactor(core): split gcs storage backend helpers
